Improve location handling of repeaters, including geocoding (#523)
The important aspect of this change is that there's now a set of inputs for the location that are kept as is and those are processed into a set of outputs. This is because some repeaters come with coordinates, some with address, some with grid, and we want to calculate all others but keep track of which ones are the original inputs.

This resolves #473 and resolves #472.